requirements.txt
# pip-requirements
# Data fetching and parsing
feedparser>=6.0.0
requests>=2.31.0
yfinance>=0.2.0
pytrends>=4.9.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Machine learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
transformers>=4.30.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# Google Sheets integration
gspread>=5.12.0
google-auth>=2.23.0

# Deep learning backend for transformers
torch>=2.0.0
